+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @layerswap/utils
 
+## 2.0.1
+
+### Patch Changes
+
+- Wallet connection readiness fixes
+- Updated dependencies
+  - @layerswap/widget-types@1.0.1
+
 ## 2.0.0
 
 ### Major Changes